refactor: move _utils, _types outside of app
