+++ src/types/hub.ts
@@ -1,5 +1,5 @@
 /**
- * File:     _types/Hub.ts
+ * File:     types/Hub.ts
  * Purpose:  Types for the Hub.
  * Authors:  Ojos Project & Iris contributors
  * License:  GNU General Public License v3.0
